backend/requirements.txt
passlib~=1.7.4
tortoise-orm~=0.19.3
pytest~=7.4.0
fastapi~=0.100.0
httpx~=0.24.1
starlette~=0.27.0
PyJWT~=2.7.0
pydantic~=1.10.7
python-dotenv~=0.21.1
asyncpg~=0.27.0
uvicorn~=0.20
python-multipart~=0.0.5
